Polyfill for casper-storage
