+++ index.js
@@ -3,8 +3,10 @@
  */
 import 'react-native-gesture-handler';
 import './polyfill';
+import './shim';
+import './shim-legacy';
+
 import { AppRegistry, LogBox } from 'react-native';
-import './shim'; // Import polyfill features
 import App from './App';
 import { name as appName } from './app.json';
 LogBox.ignoreLogs(['new NativeEventEmitter']);
