+++ package.json
@@ -26,14 +26,15 @@
     "fastlane:beta": "yarn env:beta && cd ios/fastlane && fastlane beta",
     "fastlane:release": "yarn env:mainnet && cd ios/fastlane && fastlane beta",
     "fastlane:beta:android": "yarn env:beta && cd android/fastlane && fastlane internal",
-    "fastlane:release:android": "yarn env:mainnet && cd android/fastlane && fastlane internal --verbose"
+    "fastlane:release:android": "yarn env:mainnet && cd android/fastlane && fastlane internal --verbose",
+    "postinstall": "rn-nodeify --install process,stream,crypto --hack --yarn"
   },
   "dependencies": {
     "@ethersproject/bignumber": "^5.5.0",
     "@ledgerhq/hw-app-eth": "^6.23.0",
     "@ledgerhq/react-native-hw-transport-ble": "^6.20.0",
     "@popmotion/popcorn": "^0.4.4",
-    "@react-native-async-storage/async-storage": "^1.15.8",
+    "@react-native-async-storage/async-storage": "1.17.3",
     "@react-native-clipboard/clipboard": "^1.8.4",
     "@react-native-community/blur": "^3.6.0",
     "@react-native-community/cameraroll": "^4.1.2",
@@ -47,13 +48,14 @@
     "@types/qs": "^6.9.7",
     "@types/redux-logger": "^3.0.8",
     "@zondax/ledger-casper": "^0.0.1",
+    "add": "2.0.6",
     "appcenter-cli": "^2.10.0",
     "axios": "^0.21.1",
     "base-64": "^1.0.0",
     "big-integer": "1.6.51",
     "big.js": "^6.1.1",
     "buffer": "^6.0.3",
-    "casper-storage": "^0.2.1",
+    "casper-storage": "0.2.3",
     "casperdash-js-sdk": "^0.0.1",
     "d3-shape": "^3.0.1",
     "formik": "^2.2.9",
@@ -63,6 +65,7 @@
     "lodash": "^4.17.21",
     "memoize-one": "^6.0.0",
     "moment": "^2.29.1",
+    "process": "^0.11.0",
     "prop-types": "^15.7.2",
     "randomcolor": "^0.6.2",
     "react": "17.0.2",
@@ -75,6 +78,7 @@
     "react-native-bluetooth-state-manager": "^1.3.2",
     "react-native-camera": "^4.2.1",
     "react-native-col": "^1.1.5",
+    "react-native-crypto": "^2.1.0",
     "react-native-device-info": "^8.1.4",
     "react-native-easy-toast": "^2.0.0",
     "react-native-elements": "^3.4.2",
@@ -114,6 +118,7 @@
     "react-native-video": "^5.1.1",
     "react-native-webview": "^11.6.5",
     "react-redux": "^7.2.4",
+    "readable-stream": "1.0.33",
     "redux": "^4.1.0",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
@@ -122,6 +127,7 @@
     "rn-fetch-blob": "^0.12.0",
     "rn-material-ui-textfield": "^1.0.5",
     "rxjs": "^7.5.2",
+    "stream-browserify": "^1.0.0",
     "text-encoding": "^0.7.0",
     "victory-native": "^36.3.1",
     "yup": "^0.32.9"
@@ -145,9 +151,28 @@
     "metro-react-native-babel-preset": "^0.66.2",
     "react-native-svg-transformer": "^0.14.3",
     "react-test-renderer": "17.0.2",
+    "rn-nodeify": "10.3.0",
     "typescript": "^4.3.5"
   },
   "jest": {
     "preset": "react-native"
+  },
+  "react-native": {
+    "crypto": "react-native-crypto",
+    "_stream_transform": "readable-stream/transform",
+    "_stream_readable": "readable-stream/readable",
+    "_stream_writable": "readable-stream/writable",
+    "_stream_duplex": "readable-stream/duplex",
+    "_stream_passthrough": "readable-stream/passthrough",
+    "stream": "stream-browserify"
+  },
+  "browser": {
+    "crypto": "react-native-crypto",
+    "_stream_transform": "readable-stream/transform",
+    "_stream_readable": "readable-stream/readable",
+    "_stream_writable": "readable-stream/writable",
+    "_stream_duplex": "readable-stream/duplex",
+    "_stream_passthrough": "readable-stream/passthrough",
+    "stream": "stream-browserify"
   }
 }
